ImageSliderItem and Image Refactor
