+++ src/component/App.css
@@ -1,4 +1,14 @@
-.App {
+h1 {
+  font-family: 'Megrim', cursive;
+}
+
+
+
+
+
+
+
+/* .App {
   text-align: center;
 }
 
@@ -19,4 +29,4 @@
 
 .App-link {
   color: #09d3ac;
-}
+} */
