+++ src/component/ImageSlider.css
@@ -1,0 +1,13 @@
+.slider {
+  display: flex;
+  flex-flow: row wrap;
+  width: 380px;
+  justify-content: space-between;
+}
+.slider__arrow{
+  display: block;
+  align-self: center;
+}
+.slider__image {
+  padding: 0px;
+}
