--- src/data.js
@@ -1,11 +1,0 @@
-export default [
-    "./images/bork.png",
-    "./images/coffee-mittens.png",
-    "./images/snow-bubble.png",
-    "./images/icesicles.png",
-    "./images/person-in-snow.png",
-    "./images/snow-bubble-2.png",
-    "./images/snowflake.png",
-    "./images/tweet.png",
-    "./images/winter-scene.png",
-];
