+++ src/data/images.js
@@ -1,0 +1,11 @@
+export default [
+    "./images/tiles/bork.png",
+    "./images/tiles/coffee-mittens.png",
+    "./images/tiles/snow-bubble.png",
+    "./images/tiles/icesicles.png",
+    "./images/tiles/person-in-snow.png",
+    "./images/tiles/snow-bubble-2.png",
+    "./images/tiles/snowflake.png",
+    "./images/tiles/tweet.png",
+    "./images/tiles/winter-scene.png",
+];
